requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
tortoise-orm==0.19.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0.post2
geopy==2.3.0
numpy==1.24.3
scikit-learn==1.2.2
redis==4.5.5
aioredis==2.0.1
celery==5.2.7
python-dotenv==1.0.0
bcrypt==4.0.1
pandas==2.0.1
matplotlib==3.7.1
aiofiles==23.1.0 